User authentication with firebase (#44)
* chore: request endpoint for getting etiquette votes

* chore: request endpoints for posting and patching etiquette votes

* chore: update request

* build: firebase authentication, user reg, login, logout
Added endpoints for registration, login and logout. This uses firebase to authenticate the token that is received at the front end, and set a session cookie.
Note that token and session cookie are different things in firebase, so to be properly logged in we need the session cookie. The token is just to authenticate the uid
generated by firebase.
Also added middleware to verify the session cookie. This can be used to protect routes, but I haven't applied it to any routes yet.
Need to set FIREBASE_SERVICE_ACCOUNT_KEY=ewogICJ0eXBlIjog... in .env file
Since Andy has these details he might need to do this on the live backend or change the person managing firebase.
I did not write prisma database updates for user details because I don't know how. This can be done in authModel.ts in auth -> models folder.

* fix: express types version

* fix: express types resolutions to use latest

* fix: temporary typescript error fixes

* fix: authetication with firebase typescript errors

---------

Co-authored-by: Brian <71677184+brianaqp@users.noreply.github.com>